rewrote layout fetching code
layouts are now be loaded from a package by default, falling back on
loading them from GitHub. Additionally, loading from GitHub or a local
directory can be requested with a command line switch.

This saves requests when the package is installed, does the right thing
when using handkerchief as a standalone switch, and makes development
from a local directory possible.